refactor(proxy): split attachment client methods and add transport tests
Move meeting and past-meeting attachment ITX calls into dedicated files
and cover shared HTTP error mapping with unit tests.
